+++ tests/Rules/NullableTest.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Somnambulist\Components\Validation\Tests;
+namespace Somnambulist\Components\Validation\Tests\Rules;
 
 use PHPUnit\Framework\TestCase;
 use Somnambulist\Components\Validation\Factory;
@@ -10,7 +10,7 @@ use const UPLOAD_ERR_OK;
  * Class NullableTest
  *
  * @package    Somnambulist\Components\Validation\Tests
- * @subpackage Somnambulist\Components\Validation\Tests\NullableTest
+ * @subpackage Somnambulist\Components\Validation\Tests\Rules\NullableTest
  */
 class NullableTest extends TestCase
 {