+++ tests/Rules/NumbersTest.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Somnambulist\Components\Validation\Tests;
+namespace Somnambulist\Components\Validation\Tests\Rules;
 
 use PHPUnit\Framework\TestCase;
 use Somnambulist\Components\Validation\Factory;
@@ -9,7 +9,7 @@ use Somnambulist\Components\Validation\Factory;
  * Class NumbersTest
  *
  * @package    Somnambulist\Components\Validation\Tests
- * @subpackage Somnambulist\Components\Validation\Tests\NumbersTest
+ * @subpackage Somnambulist\Components\Validation\Tests\Rules\NumbersTest
  */
 class NumbersTest extends TestCase
 {